+++ secrets_env/commands/completion.py
@@ -1,10 +1,10 @@
-import logging
 import os
 import sys
 import warnings
 
 import click
 
+import secrets_env.utils
 from secrets_env.commands.core import entrypoint, with_output_options
 
 
@@ -19,8 +19,6 @@ def completion(ctx: click.Context, shell: str):
 
        eval "$(secrets.env completion)"
     """
-    logger = logging.getLogger(__name__)
-
     # get prog name
     prog_name = os.path.basename(sys.argv[0])
     if prog_name.lower().endswith(".py"):
@@ -32,8 +30,7 @@ def completion(ctx: click.Context, shell: str):
 
     # get shell
     if not shell:
-        shell = os.path.basename(os.environ["SHELL"])
-        logger.debug("Detect %s", shell)
+        shell, _ = secrets_env.utils.detect_shell()
 
     # print script
     import click.shell_completion