+++ secrets_env/realms/poetry/cleo.py
@@ -75,6 +75,7 @@ class CleoHandler(logging.Handler):
             self.handleError(record)
             return
 
+        verbosity = typing.cast(Verbosity, verbosity)  # to satisfy pyright
         self.output.write_line(msg, verbosity=verbosity)