+++ secrets_env/utils.py
@@ -225,3 +225,35 @@ def _show_warning(
         s = warnings.formatwarning(message, category, filename, lineno, line)
         logger = logging.getLogger("py.warnings")
         logger.warning(str(s))
+
+
+def detect_shell() -> tuple[str, Path]:
+    # try to detect shell via shellingham
+    if pair := _detect_shell_via_shellingham():
+        return pair
+
+    # fallback to default
+    if os.name == "posix":
+        raw_shell_path = os.getenv("SHELL")
+    elif os.name == "nt":
+        raw_shell_path = os.getenv("COMSPEC")
+    if not raw_shell_path:
+        raise NotImplementedError(f"OS {os.name!r} support not available")
+
+    shell_path = Path(raw_shell_path)
+    logger.debug("Use current shell: %s", shell_path)
+    return shell_path.stem.lower(), shell_path
+
+
+def _detect_shell_via_shellingham() -> tuple[str, Path] | None:
+    try:
+        import shellingham
+    except ImportError:
+        return None
+
+    try:
+        shell, path = shellingham.detect_shell()
+        logger.debug("Detect shell: %s", path)
+        return shell, Path(path)
+    except shellingham.ShellDetectionFailure:
+        return None